:art: 重新生成coupon

+++ codingce-gulimall/gulimall-coupon/src/main/java/cn/com/codingce/coupon/entity/SeckillSkuRelationEntity.java
@@ -13,7 +13,7 @@ import lombok.Data;
  * 
  * @author mxz
  * @email codingce@gmail.com
- * @date 2021-01-29 16:27:43
+ * @date 2021-02-24 12:31:20
  */
 @Data
 @TableName("sms_seckill_sku_relation")
@@ -44,11 +44,11 @@ public class SeckillSkuRelationEntity implements Serializable {
 	/**
 	 * 秒杀总量
 	 */
-	private BigDecimal seckillCount;
+	private Integer seckillCount;
 	/**
 	 * 每人限购数量
 	 */
-	private BigDecimal seckillLimit;
+	private Integer seckillLimit;
 	/**
 	 * 排序
 	 */
